refactor out the useGrid hook
