+++ windwatts-ui/src/hooks/useDownloadCSVFile.ts
@@ -1,11 +1,6 @@
 import { useState, useContext, useMemo } from "react";
 import useSWR from "swr";
-import {
-  getExportCSV,
-  getBatchExportCSV,
-  getNearestGridLocation,
-  getModelInfo,
-} from "../services/api";
+import { getExportCSV, getBatchExportCSV, getModelInfo } from "../services/api";
 import { downloadWindDataCSV, downloadWindDataZIP } from "../services/download";
 import { SettingsContext } from "../providers/SettingsContext";
 import { GridLocation } from "../types";
@@ -53,7 +48,7 @@ export const useDownloadCSVFile = () => {
     try {
       setIsDownloading(true);
 
-      const response = await getExportCSV(
+      const response: Response = await getExportCSV(
         {
           gridIndex: gridIndex,
           dataModel: dataModel!,
@@ -78,7 +73,7 @@ export const useDownloadCSVFile = () => {
     try {
       setIsDownloading(true);
 
-      const response = await getBatchExportCSV(
+      const response: Response = await getBatchExportCSV(
         {
           gridLocations: gridLocations,
           dataModel: dataModel!,
@@ -113,63 +108,5 @@ export const useDownloadCSVFile = () => {
     setYearSet,
     fullYearRange,
     sampleYearRange,
-  };
-};
-
-export const useNearestGridLocation = (n_neighbors: number = 1) => {
-  const { currentPosition, preferredModel } = useContext(SettingsContext);
-  const dataModel =
-    preferredModel === "ensemble-quantiles" ? "era5-quantiles" : preferredModel;
-  const { lat, lng } = currentPosition || {};
-
-  const shouldFetch = lat && lng && dataModel;
-
-  // Memoize the SWR key to prevent unnecessary re-renders
-  const swrKey = useMemo(() => {
-    if (!shouldFetch) return null;
-    return JSON.stringify({
-      lat,
-      lng,
-      n_neighbors,
-      dataModel,
-      type: "nearest-grid",
-    });
-  }, [shouldFetch, lat, lng, n_neighbors, dataModel]);
-
-  const {
-    isLoading,
-    data,
-    error,
-    mutate: retry,
-  } = useSWR(
-    swrKey,
-    () =>
-      getNearestGridLocation({
-        lat: lat!,
-        lng: lng!,
-        n_neighbors: n_neighbors,
-        dataModel: dataModel!,
-      }),
-    {
-      revalidateOnFocus: false,
-      revalidateOnReconnect: false,
-      revalidateIfStale: false,
-      dedupingInterval: 60000, // Cache for 1 minute
-    }
-  );
-
-  const gridLocations =
-    data?.locations?.map((location: GridLocation) => ({
-      latitude: location.latitude,
-      longitude: location.longitude,
-      index: location.index,
-    })) || [];
-
-  return {
-    gridLocations,
-    isLoading,
-    error: error?.message || error || null,
-    hasData: gridLocations.length > 0,
-    retry,
   };
 };
